+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+## [2.3.12]
 ### enhancements
 - add new ways of customization for theme settings